requirements.txt
Django==2.2.7
django-extensions==2.2.5
djangorestframework==3.10.3
psycopg2==2.8.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0